requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==3.0.10
mysqlclient==2.2.4
python-dotenv==0.19.0
networkx==2.6.3
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
requests==2.26.0
python-jose==3.3.0
googlemaps==4.10.0
gunicorn==20.1.0
pytest==6.2.5
black==21.7b0
flake8==3.9.2
.